3.7.5 Further document multiexp functions
